Feat/RAMDisk

+++ morphct/run_MorphCT.py
@@ -241,10 +241,16 @@ class simulation:
 
     def make_dir_tree(self):
         print("Sorting out directory structure...")
-        # Delete the ORCA information if the user has asked to
-        if self.overwrite_current_data is True:
+        # Delete the ORCA information if the user has asked to (as long as it's not in
+        # the same place as the morphology data!!)
+        if (self.overwrite_current_data is True) and (
+            self.output_orca_directory != self.output_morphology_directory
+        ):
             print("OVERWRITE CURRENT DATA IS TRUE...EMPTYING ORCA DIR")
-            shutil.rmtree(self.output_orca_directory)
+            try:
+                shutil.rmtree(self.output_orca_directory)
+            except FileNotFoundError:
+                print("Directory already empty. Continuing...")
         # NOTE: Don't delete the morphology directory - sometimes we just want to
         # update and recalculate things, not start the whole pipeline again.
         # Delete any previous data if the user asked to
